Sanitize resource names
This is implemented to save resources on the disk safely, keeping the filenames close to original.
